+++ storefront/src/lib/data/orders.ts
@@ -23,7 +23,7 @@ export const retrieveOrder = async (id: string) => {
       method: "GET",
       query: {
         fields:
-          "*payment_collections.payments,*items,*items.metadata,*items.variant,*items.product,+items.product.food_details.*,+items.product.wine_details.*",
+          "*payment_collections.payments,*items,*items.metadata,*items.variant,*items.product,+items.product.food_details.*,+items.product.wine_details.*,+metadata",
       },
       headers,
       next,
@@ -53,7 +53,8 @@ export const listOrders = async (
         limit,
         offset,
         order: "-created_at",
-        fields: "*items,+items.metadata,*items.variant,*items.product",
+        fields:
+          "*items,+items.metadata,*items.variant,*items.product,+items.product.food_details.id,+metadata",
         ...filters,
       },
       headers,