+++ storefront/src/modules/order/templates/order-details-template.tsx
@@ -7,6 +7,7 @@ import ReorderButton from "@modules/account/components/reorder-button"
 import Help from "@modules/order/components/help"
 import Items from "@modules/order/components/items"
 import OrderDetails from "@modules/order/components/order-details"
+import FoodOrderStatus from "@modules/order/components/food-order-status"
 import OrderSummary from "@modules/order/components/order-summary"
 import ShippingDetails from "@modules/order/components/shipping-details"
 import React from "react"
@@ -45,6 +46,7 @@ const OrderDetailsTemplate: React.FC<OrderDetailsTemplateProps> = ({
         data-testid="order-details-container"
       >
         <OrderDetails order={order} showStatus />
+        <FoodOrderStatus order={order} />
         <Items order={order} />
         <ShippingDetails order={order} />
         <OrderSummary order={order} />